change: change unit size of bitmap from 32 to 64

+++ mison.go
@@ -6,7 +6,7 @@ RemoveRightmost1 removes the rightmost 1 in x.
 E.g.
 	11101000 -> 11100000
 */
-func RemoveRightmost1(x uint64) uint64 {
+func RemoveRightmost1(x uint32) uint32 {
 	return x & (x - 1)
 }
 
@@ -16,7 +16,7 @@ ExtractRightmost1 extract the rightmost 1 in x.
 E.g.
 	11101000 -> 00001000
 */
-func ExtractRightmost1(x uint64) uint64 {
+func ExtractRightmost1(x uint32) uint32 {
 	return x & -x
 }
 
@@ -26,6 +26,6 @@ SmearRightmost1 extract the rightmost 1 in x and smear it to the right.
 E.g.
 	11101000 -> 00001111
 */
-func SmearRightmost1(x uint64) uint64 {
+func SmearRightmost1(x uint32) uint32 {
 	return x ^ (x - 1)
 }
